+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 
 *Internal*
 * Git URL model changed inside the IR
+* `orq up` will now configure Ray's Plasma directory
 
 
 *Docs*